requirements.txt
Django==5.1.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.3.1
yara_python==4.5.1
celery==5.4.0
redis==5.2.1
kombu==5.4.2
gunicorn==23.0.0
django-cors-headers==4.7.0